Module aws_sdk_shield::error

Modules

access_denied_exception

See AccessDeniedException

access_denied_for_dependency_exception

See AccessDeniedForDependencyException

internal_error_exception

See InternalErrorException

invalid_operation_exception

See InvalidOperationException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_resource_exception

See InvalidResourceException

limits_exceeded_exception

See LimitsExceededException

locked_subscription_exception

See LockedSubscriptionException

no_associated_role_exception

See NoAssociatedRoleException

optimistic_lock_exception

See OptimisticLockException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

Structs

AccessDeniedException

Exception that indicates the specified AttackId does not exist, or the requester does not have the appropriate permissions to access the AttackId.

AccessDeniedForDependencyException

In order to grant the necessary access to the Shield Response Team (SRT) the user submitting the request must have the iam:PassRole permission. This error indicates the user did not have the appropriate permissions. For more information, see Granting a User Permissions to Pass a Role to an Amazon Web Services Service.

AssociateDRTLogBucketError

AssociateDRTRoleError

AssociateHealthCheckError

AssociateProactiveEngagementDetailsError

CreateProtectionError

CreateProtectionGroupError

CreateSubscriptionError

DeleteProtectionError

DeleteProtectionGroupError

DeleteSubscriptionError

DescribeAttackError

DescribeAttackStatisticsError

DescribeDRTAccessError

DescribeEmergencyContactSettingsError

DescribeProtectionError

DescribeProtectionGroupError

DescribeSubscriptionError

DisableProactiveEngagementError

DisassociateDRTLogBucketError

DisassociateDRTRoleError

DisassociateHealthCheckError

EnableProactiveEngagementError

GetSubscriptionStateError

InternalErrorException

Exception that indicates that a problem occurred with the service infrastructure. You

InvalidOperationException

Exception that indicates that the operation would not cause any change to occur.

InvalidPaginationTokenException

Exception that indicates that the NextToken specified in the request is invalid. Submit the request using the NextToken value that was returned in the response.

InvalidParameterException

Exception that indicates that the parameters passed to the API are invalid. If available, this exception includes details in additional properties.

InvalidResourceException

Exception that indicates that the resource is invalid. You might not have access to the resource, or the resource might not exist.

LimitsExceededException

Exception that indicates that the operation would exceed a limit.

ListAttacksError

ListProtectionGroupsError

ListProtectionsError

ListResourcesInProtectionGroupError

ListTagsForResourceError

LockedSubscriptionException

You are trying to update a subscription that has not yet completed the 1-year commitment. You can change the AutoRenew parameter during the last 30 days of your subscription. This exception indicates that you are attempting to change AutoRenew prior to that period.

NoAssociatedRoleException

The ARN of the role that you specifed does not exist.

OptimisticLockException

Exception that indicates that the resource state has been modified by another

ResourceAlreadyExistsException

Exception indicating the specified resource already exists. If available, this exception includes details in additional properties.

ResourceNotFoundException

Exception indicating the specified resource does not exist. If available, this exception includes details in additional properties.

TagResourceError

UntagResourceError

UpdateEmergencyContactSettingsError

UpdateProtectionGroupError

UpdateSubscriptionError

Enums

AssociateDRTLogBucketErrorKind

AssociateDRTRoleErrorKind

AssociateHealthCheckErrorKind

AssociateProactiveEngagementDetailsErrorKind

CreateProtectionErrorKind

CreateProtectionGroupErrorKind

CreateSubscriptionErrorKind

DeleteProtectionErrorKind

DeleteProtectionGroupErrorKind

DeleteSubscriptionErrorKind

DescribeAttackErrorKind

DescribeAttackStatisticsErrorKind

DescribeDRTAccessErrorKind

DescribeEmergencyContactSettingsErrorKind

DescribeProtectionErrorKind

DescribeProtectionGroupErrorKind

DescribeSubscriptionErrorKind

DisableProactiveEngagementErrorKind

DisassociateDRTLogBucketErrorKind

DisassociateDRTRoleErrorKind

DisassociateHealthCheckErrorKind

EnableProactiveEngagementErrorKind

GetSubscriptionStateErrorKind

ListAttacksErrorKind

ListProtectionGroupsErrorKind

ListProtectionsErrorKind

ListResourcesInProtectionGroupErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateEmergencyContactSettingsErrorKind

UpdateProtectionGroupErrorKind

UpdateSubscriptionErrorKind