Module aws_sdk_shield::model

Modules

attack_detail

See AttackDetail

attack_property

See AttackProperty

attack_statistics_data_item

See AttackStatisticsDataItem

attack_summary

See AttackSummary

attack_vector_description

See AttackVectorDescription

attack_volume

See AttackVolume

attack_volume_statistics

See AttackVolumeStatistics

contributor

See Contributor

emergency_contact

See EmergencyContact

limit

See Limit

mitigation

See Mitigation

protection

See Protection

protection_group

See ProtectionGroup

protection_group_arbitrary_pattern_limits

See ProtectionGroupArbitraryPatternLimits

protection_group_limits

See ProtectionGroupLimits

protection_group_pattern_type_limits

See ProtectionGroupPatternTypeLimits

protection_limits

See ProtectionLimits

sub_resource_summary

See SubResourceSummary

subscription

See Subscription

subscription_limits

See SubscriptionLimits

summarized_attack_vector

See SummarizedAttackVector

summarized_counter

See SummarizedCounter

tag

See Tag

time_range

See TimeRange

validation_exception_field

See ValidationExceptionField

Structs

AttackDetail

The details of a DDoS attack.

AttackProperty

Details of a Shield event. This is provided as part of an AttackDetail.

AttackStatisticsDataItem

A single attack statistics data record. This is returned by DescribeAttackStatistics along with a time range indicating the time period that the attack statistics apply to.

AttackSummary

Summarizes all DDoS attacks for a specified time period.

AttackVectorDescription

Describes the attack.

AttackVolume

Information about the volume of attacks during the time period, included in an AttackStatisticsDataItem. If the accompanying AttackCount in the statistics object is zero, this setting might be empty.

AttackVolumeStatistics

Statistics objects for the various data types in AttackVolume.

Contributor

A contributor to the attack and their contribution.

EmergencyContact

Contact information that the SRT can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

Limit

Specifies how many protections of a given type you can create.

Mitigation

The mitigation applied to a DDoS attack.

Protection

An object that represents a resource that is under DDoS protection.

ProtectionGroup

A grouping of protected resources that you and Shield Advanced can monitor as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

ProtectionGroupArbitraryPatternLimits

Limits settings on protection groups with arbitrary pattern type.

ProtectionGroupLimits

Limits settings on protection groups for your subscription.

ProtectionGroupPatternTypeLimits

Limits settings by pattern type in the protection groups for your subscription.

ProtectionLimits

Limits settings on protections for your subscription.

SubResourceSummary

The attack information for the specified SubResource.

Subscription

Information about the Shield Advanced subscription for an account.

SubscriptionLimits

Limits settings for your subscription.

SummarizedAttackVector

A summary of information about the attack.

SummarizedCounter

The counter that describes a DDoS attack.

Tag

A tag associated with an Amazon Web Services resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing or other management. Typically, the tag key represents a category, such as "environment", and the tag value represents a specific value within that category, such as "test," "development," or "production". Or you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each Amazon Web Services resource, up to 50 tags for a resource.

TimeRange

The time range.

ValidationExceptionField

Provides information about a particular parameter passed inside a request that resulted in an exception.

Enums

AttackLayer

AttackPropertyIdentifier

AutoRenew

ProactiveEngagementStatus

ProtectedResourceType

ProtectionGroupAggregation

ProtectionGroupPattern

SubResourceType

SubscriptionState

Unit

ValidationExceptionReason