Module aws_sdk_shield::operation

Structs

AssociateDRTLogBucket

Authorizes the Shield Response Team (SRT) to access the specified Amazon S3 bucket containing log data such as Application Load Balancer access logs, CloudFront logs, or logs from third party sources. You can associate up to 10 Amazon S3 buckets with your subscription.

AssociateDRTRole

Authorizes the Shield Response Team (SRT) using the specified role, to access your Amazon Web Services account to assist with DDoS attack mitigation during potential attacks. This enables the SRT to inspect your WAF configuration and create or update WAF rules and web ACLs.

AssociateHealthCheck

Adds health-based detection to the Shield Advanced protection for a resource. Shield Advanced health-based detection uses the health of your Amazon Web Services resource to improve responsiveness and accuracy in attack detection and mitigation.

AssociateProactiveEngagementDetails

Initializes proactive engagement and sets the list of contacts for the Shield Response Team (SRT) to use. You must provide at least one phone number in the emergency contact list.

CreateProtection

Enables Shield Advanced for a specific Amazon Web Services resource. The resource can be an Amazon CloudFront distribution, Elastic Load Balancing load balancer, Global Accelerator accelerator, Elastic IP Address, or an Amazon Route 53 hosted zone.

CreateProtectionGroup

Creates a grouping of protected resources so they can be handled as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

CreateSubscription

Activates Shield Advanced for an account.

DeleteProtection

Deletes an Shield Advanced Protection.

DeleteProtectionGroup

Removes the specified protection group.

DeleteSubscription

Removes Shield Advanced from an account. Shield Advanced requires a 1-year subscription commitment. You cannot delete a subscription prior to the completion of that commitment.

DescribeAttack

Describes the details of a DDoS attack.

DescribeAttackStatistics

Provides information about the number and type of attacks Shield has detected in the last year for all resources that belong to your account, regardless of whether you've defined Shield protections for them. This operation is available to Shield customers as well as to Shield Advanced customers.

DescribeDRTAccess

Returns the current role and list of Amazon S3 log buckets used by the Shield Response Team (SRT) to access your Amazon Web Services account while assisting with attack mitigation.

DescribeEmergencyContactSettings

A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

DescribeProtection

Lists the details of a Protection object.

DescribeProtectionGroup

Returns the specification for the specified protection group.

DescribeSubscription

Provides details about the Shield Advanced subscription for an account.

DisableProactiveEngagement

Removes authorization from the Shield Response Team (SRT) to notify contacts about escalations to the SRT and to initiate proactive customer support.

DisassociateDRTLogBucket

Removes the Shield Response Team's (SRT) access to the specified Amazon S3 bucket containing the logs that you shared previously.

DisassociateDRTRole

Removes the Shield Response Team's (SRT) access to your Amazon Web Services account.

DisassociateHealthCheck

Removes health-based detection from the Shield Advanced protection for a resource. Shield Advanced health-based detection uses the health of your Amazon Web Services resource to improve responsiveness and accuracy in attack detection and mitigation.

EnableProactiveEngagement

Authorizes the Shield Response Team (SRT) to use email and phone to notify contacts about escalations to the SRT and to initiate proactive customer support.

GetSubscriptionState

Returns the SubscriptionState, either Active or Inactive.

ListAttacks

Returns all ongoing DDoS attacks or all DDoS attacks during a specified time

ListProtectionGroups

Retrieves the ProtectionGroup objects for the account.

ListProtections

Lists all Protection objects for the account.

ListResourcesInProtectionGroup

Retrieves the resources that are included in the protection group.

ListTagsForResource

Gets information about Amazon Web Services tags for a specified Amazon Resource Name (ARN) in Shield.

TagResource

Adds or updates tags for a resource in Shield.

UntagResource

Removes tags from a resource in Shield.

UpdateEmergencyContactSettings

Updates the details of the list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

UpdateProtectionGroup

Updates an existing protection group. A protection group is a grouping of protected resources so they can be handled as a collective. This resource grouping improves the accuracy of detection and reduces false positives.

UpdateSubscription

Updates the details of an existing subscription. Only enter values for parameters you want to change. Empty parameters are not updated.